Add first unit tests for JobService
